Add files via upload
A simple GUI version of the number guessing game written in standard python language. The game keeps track of attempts and displays the guess counter for each iteration of play.

NOTE: The file breezypythongui.py MUST be in the same directory as this file for the app to run correctly!